Redesigned block copying mechanism again. Now we're using the bitwise type checker.

+++ include/stack.c
@@ -31,7 +31,7 @@ struct cognate_stack
 {
   cognate_list items;   // The list holding the stack itself.
   ptrdiff_t    size;    // Allocated size of the stack.
-  long uncopied_blocks; // Number of cognate_blocks this function has produced (or consumed).
+  size_t uncopied_blocks; // Number of cognate_blocks this function has produced (or consumed).
 };
 
 typedef struct cognate_stack cognate_stack;
